requirements.txt
pyyaml>=5.3.1
setuptools>=65.6.3